Updated map with new spawn

+++ src/main/java/com/hugman/the_towers/config/TheTowersConfig.java
@@ -8,13 +8,13 @@ import xyz.nucleoid.plasmid.game.common.team.GameTeam;
 
 import java.util.List;
 
-public record TheTowersConfig(PlayerConfig playerConfig, List<GameTeam> teamConfig, Identifier mapTemplateId, int maxHealth, boolean healthStealth, long respawnCooldown) {
+public record TheTowersConfig(PlayerConfig playerConfig, List<GameTeam> teamConfig, Identifier mapTemplateId, int maxHealth, boolean healthStealth, int respawnCooldown) {
 	public static final Codec<TheTowersConfig> CODEC = RecordCodecBuilder.create(instance -> instance.group(
 			PlayerConfig.CODEC.fieldOf("players").forGetter(TheTowersConfig::playerConfig),
 			GameTeam.CODEC.listOf().fieldOf("teams").forGetter(TheTowersConfig::teamConfig),
 			Identifier.CODEC.fieldOf("map").forGetter(TheTowersConfig::mapTemplateId),
 			Codec.INT.fieldOf("max_health").forGetter(TheTowersConfig::maxHealth),
 			Codec.BOOL.fieldOf("health_stealth").forGetter(TheTowersConfig::healthStealth),
-			Codec.LONG.optionalFieldOf("respawn_cooldown", 5L).forGetter(TheTowersConfig::respawnCooldown)
+			Codec.INT.optionalFieldOf("respawn_cooldown", 5).forGetter(TheTowersConfig::respawnCooldown)
 	).apply(instance, TheTowersConfig::new));
 }
